Handling up to 8 clients
Sending packets is now done via a struct shared between the client and the server.

Each client has a vector of x elements, each a square of a different colour. When a client connects, it gives the client an ID, which is used to determine which square the player is and maintain it across all clients connected to the server.